Some documentation and a minor conversion fix.

+++ sysinfo.go
@@ -8,7 +8,7 @@ import (
 )
 
 type Sysinfo struct {
-	uptime    int64
+	uptime    uint64
 	memTotal  uint64
 	memFree   uint64
 	memShared uint64
@@ -16,9 +16,10 @@ type Sysinfo struct {
 	swapTotal uint64
 	swapFree  uint64
 	load      [3]uint64
-	procs     uint16
+	procs     uint64
 }
 
+// Update updates the Sysinfo struct with current system information.
 func (s *Sysinfo) Update() {
 	info := syscall.Sysinfo_t{}
 	err := syscall.Sysinfo(&info)
@@ -27,7 +28,7 @@ func (s *Sysinfo) Update() {
 		return
 	}
 
-	s.uptime = info.Uptime
+	s.uptime = uint64(info.Uptime)
 	s.memTotal = info.Totalram
 	s.memFree = info.Freeram
 	s.memShared = info.Sharedram
@@ -35,30 +36,36 @@ func (s *Sysinfo) Update() {
 	s.swapTotal = info.Totalswap
 	s.swapFree = info.Freeswap
 	s.load = info.Loads
-	s.procs = info.Procs
+	s.procs = uint64(info.Procs)
 }
 
-func (s Sysinfo) Uptime() int64 {
+// Uptime returns system uptime.
+func (s Sysinfo) Uptime() uint64 {
 	return s.uptime
 }
 
+// MemTotal returns all available memory.
 func (s Sysinfo) MemTotal() uint64 {
 	return s.memTotal
 }
 
+// MemFree returns amount of free memory in bytes.
 func (s Sysinfo) MemFree() uint64 {
 	return s.memFree + s.memBuffer
 }
 
+// MemUsed returns amount of used memory in bytes.
 func (s Sysinfo) MemUsed() uint64 {
 	return s.MemTotal() - s.MemFree()
 }
 
+// MemUsedPercent returns used memory in percent.
 func (s Sysinfo) MemUsedPercent() uint64 {
 	return s.MemUsed() * 100 / s.MemTotal()
 }
 
-func (s Sysinfo) Procs() uint16 {
+// Procs returns number of currently running processes.
+func (s Sysinfo) Procs() uint64 {
 	return s.procs
 }
 
@@ -67,6 +74,7 @@ func (s Sysinfo) Load() uint64 {
 	return s.load[1]
 }
 
+// NewSysinfo returns populated Sysinfo struct.
 func NewSysinfo() *Sysinfo {
 	s := &Sysinfo{}
 	s.Update()
